# ARROW-8553: [C++] Optimize unaligned bitmap operations