Unaligned bitmap operations are currently calculated bit-by-bit.
Processing it word-by-word in uint64 improves performance significantly.
BenchmarkBitmapAnd/32768/1 jumps from 56M/s to 4.1G/s on my test machine.

This patch also improves unit test to cover new code thoroughly.